Criando endpoints de orçamento
Criando endpoints de orçamento e logica de construção do mesmo

+++ src/main/java/br/com/cvc/hotels/dto/PriceDTO.java
@@ -10,8 +10,8 @@ import lombok.Data;
 public class PriceDTO {
 
 	@Expose
-	private int adult;
+	private double adult;
 	
 	@Expose
-	private int child;
+	private double child;
 }
